Remake exercício calc Super-homem
Refiz o exercício, acredito dessa vez ser a  correta.
