+++ felipe honorio moraes/Calculadora de Super Força/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=b32aa9d0
+build.xml.script.CRC32=0f41b3c0
+build.xml.stylesheet.CRC32=f85dc8f2@1.95.0.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=b32aa9d0
+nbproject/build-impl.xml.script.CRC32=6a233a66
+nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.95.0.48
