+++ felipe honorio moraes/Calculadora de Super Força/nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\Pichau\\AppData\\Roaming\\NetBeans\\12.0\\build.properties
