--- felipe honorio moraes/Super_Hero_Power_Calculator/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
