--- felipe honorio moraes/Super_Hero_Power_Calculator/.idea/modules.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/Super_Hero_Power_Calculator.iml" filepath="$PROJECT_DIR$/Super_Hero_Power_Calculator.iml" />
-    </modules>
-  </component>
-</project>
